refactor: replace posthog.warehouse.utils with posthog.sync
