+++ posthog/temporal/data_imports/pipelines/helpers.py
@@ -4,7 +4,7 @@ from zoneinfo import ZoneInfo
 from django.db.models import F
 
 from posthog.warehouse.types import IncrementalFieldType
-from posthog.warehouse.util import database_sync_to_async
+from posthog.sync import database_sync_to_async
 
 
 @database_sync_to_async
