+++ posthog/temporal/tests/data_modeling/test_run_workflow.py
@@ -43,7 +43,7 @@ from posthog.temporal.tests.utils.events import generate_test_events_in_clickhou
 from posthog.warehouse.models.datawarehouse_saved_query import DataWarehouseSavedQuery
 from posthog.warehouse.models.modeling import DataWarehouseModelPath
 from posthog.warehouse.models.table import DataWarehouseTable
-from posthog.warehouse.util import database_sync_to_async
+from posthog.sync import database_sync_to_async
 
 pytestmark = [pytest.mark.asyncio, pytest.mark.django_db]
