+++ posthog/warehouse/models/external_data_source.py
@@ -14,7 +14,7 @@ from posthog.models.utils import (
     UUIDModel,
     sane_repr,
 )
-from posthog.warehouse.util import database_sync_to_async
+from posthog.sync import database_sync_to_async
 
 logger = structlog.get_logger(__name__)
